In FSFS and FSX, reject commits which cause MD5 or SHA1 collisions.
Since r1785754 the rep-cache no longer depends on MD5/SHA1 hash algorithms
alone to index content. So we can store colliding content without causing
repository corruption if the rep-cache is active, which is great.

However, similar problems still exist in (at least) the RA layer and the
working copy. Until those are fixed, rejecting content which causes a hash
collision is the safest approach and avoids the undesired consequences of
storing such content.

Rejecting such content only works if rep-sharing is enabled (the default).
Content with hash collisions can still be stored if rep-sharing is disabled
but please beware of annoying consequences before doing so.

This commit is a follow-up to r1785754 and transforms a warning message
added in that revision into a fatal error.

Related dev@ discussion starts at:
https://svn.haxx.se/dev/archive-2017-05/0053.shtml
Subject: [PATCH] reject SHA1 collisions
Message-ID: <20170509132523.GN55397@ted.stsp.name>

* subversion/libsvn_fs_fs/transaction.c,
  subversion/libsvn_fs_x/transaction.c
  (get_shared_rep): Treat a checksum collision with different content
    as a fatal error.

* subversion/tests/libsvn_fs/fs-test.c
  (flag_fs_warnings): Remove, not needed anymore.
  (test_rep_sharing_strict_content_check): Adjust test expectations.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1794611 13f79535-47bb-0310-9956-ffa450edef68