Search for gpg-agent sockets in (/var)/run.
This reverts executing gpgconf to find the socket, as implemented in r1794166.
Instead, follow mostly the same procedure for socket lookup as GPG.  GPG's
hashing of non-default $GNUPGHOME to use as a portion of the path under
(/var)/run is not (yet) done, since it relies on implementing z-base-32
encoding.

* subversion/libsvn_subr/gpg_agent.c
  (find_gpgconf_agent_socket): Removed function which ran gpgconf ...
  (find_gpg_agent_socket): ... and added one that performs the various checks
   to determine where the user's gpg-agent socket resides.
  (find_running_gpg_agent): Replace the socket lookup with a call to
   find_gpg_agent_socket.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1795087 13f79535-47bb-0310-9956-ffa450edef68