Make the core of the wc commit do less by moving the logic responsible for creating temporary files with the detranslated working file contents to one of its calling sites.
The mentioned core function is svn_wc__internal_transmit_text_deltas(),
and the ability to create such files is only required for the deprecated
svn_wc_transmit_text_deltas2() and svn_wc_transmit_text_deltas() API
functions.  That's where we will move this logic.

The side effect of this change is that it also fixes a minor bug — such
temporary files could be left behind in case of some of the errors.

* subversion/libsvn_wc/wc.h
  (svn_wc__internal_transmit_text_deltas): Accept a stream to the temporary
   file, instead of an out argument where we'd (previously) put the path
   to the temporary file.

* subversion/libsvn_wc/adm_crawler.c
  (svn_wc__internal_transmit_text_deltas): Don't create and delete the
   temporary file, and only optionally wrap the passed-in stream with
   the copying_stream() handler.

* subversion/libsvn_wc/deprecated.c
  (svn_wc_transmit_text_deltas2): Create the temporary file in this
   deprecated function.  Pass the disowned tempstream to the core
   implementation, svn_wc_transmit_text_deltas2().  Properly close
   the stream and don't leave the temporary file on the disk in case
   of all errors.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1804816 13f79535-47bb-0310-9956-ffa450edef68