Make the fsfs stats test work properly with LZ4 compression.
One of the expectations has been that for our test Greek Tree repository,
the on-disk size of the directory representations should be _more_ than
their actual size (due to the overheads associated with serializing them
on the disk).

This assumption is quite sensitive to the used compression algorithm.
For example, the actual size of the directory representations is 618 bytes.
Assuming directory deltification is enabled, with zlib compression the size
is 708 bytes (more than actual), but with LZ4 compression it's 571 bytes,
and that is less than the actual size.  While I think that such result can
be explained just by the fact that the compression algorithms are just
different, the test should take that into account.

* subversion/tests/libsvn_fs_fs/fs-fs-private-test.c
  (verify_representation_stats): Expect sane relation between the packed
   and the actual sizes, but don't assume that the packed size will
   _always_ exceed the actual size.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1805896 13f79535-47bb-0310-9956-ffa450edef68