ra_serf: Prevent the server from generating and sending the full MERGE response in cases when we don't require it.
The full response is not required when working over HTTPv2 protocol.
When working over HTTPv1, it's only required when the new working copy
properties need to be stored as part of a commit (indicated by a non-null
svn_ra_push_wc_prop_func_t callback).

Disabling the full merge response is achieved by setting "no-merge-response"
in the X-SVN-Options header, and is something that was done by ra_neon,
but apparently, this optimization was not reimplemented in ra_serf.

* subversion/libsvn_ra_serf/merge.c
  (setup_merge_headers): Disable the full merge response in the majority
   of cases when it's not required.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1806017 13f79535-47bb-0310-9956-ffa450edef68