Fix issue #4700 for property changes. Coincidentally, this commit also fixes issue #4623 for FSFS.
If rep-sharing is disabled, switching from svn_fs_props_different()
to svn_fs_props_changed() fixes issue #4700 for properties, similar
to how r1813794 fixed it for file content.

However, if rep-sharing is enabled, svn_fs_props_changed() does not work
as advertised because properties do not carry a SHA1 checksum with a
"uniquifier" which identifies the transaction they were created in.
The uniquifier is used by svn_fs_props_changed() to tell apart property
representations which share content but were created in different revisions.

To fix that problem, make FSFS write SHA1 checksums along with uniquifiers
for file properties, just as it is already done for file content.

A source code comment indicates that SHA1 checksums for property reps
were not written due to concerns over disk space. In hindsight, this was
a bad trade-off because it affected correctness of svn_fs_props_changed().

* subversion/libsvn_fs_fs/transaction.c
  (svn_fs_fs__set_proplist, write_final_rev): Create uniquifiers on prop reps.

* subversion/libsvn_repos/reporter.c
  (update_entry): Use svn_fs_props_changed() instead of using
   svn_fs_props_different(). The reasoning is the same as for the
   file content fix committed in r1813794.

* basic_tests.py
  (null_prop_update_last_changed_revision): This test now passes for FSFS.
   It still fails for BDB, though.

* subversion/tests/cmdline/svnadmin_tests.py
  (verify_non_utf8_paths): Update test expectations. This test messes about
   with revision file data and thus needs an update to keep passing.
  (dump_no_op_prop_change): This test now passes as well.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1813898 13f79535-47bb-0310-9956-ffa450edef68