Viewspec: Add an experimental viewspec output command.
This reports the current layout of the working copy, in terms of switched
URLs, mixed revisions, and mixed depths.

The current CLI command is:

  svn info --viewspec [PATH]

The current output is in the form of a series of 'svn' commands which can
be used as command-line input to recreate the layout, although the behaviour
when doing so is not optimal.

Patch by: rhuijben
(tweaked by me)

* subversion/include/svn_client.h
  (svn_client_layout_func_t,
   svn_client_layout_list): New.

* subversion/libsvn_client/layout.c
  New file.

* subversion/svn/cl.h
  (svn_cl__opt_state_t): Add a 'viewspec' option.

* subversion/svn/info-cmd.c
  (layout_list_baton_t,
   layout_func,
   cl_layout_list): New.
  (svn_cl__info): Call it.

* subversion/svn/svn.c
  (svn_cl__longopt_t,
   svn_cl__options): Add a 'viewspec' option.
  (svn_cl__cmd_table): Let the 'info' command take the option.
  (sub_main): Parse the option.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1826864 13f79535-47bb-0310-9956-ffa450edef68