Make 'svnadmin verify --keep-going' always show an error summary.
Previously, the error summary was not displayed if the --quiet option
was used in combination with --keep-going. The purpose of the --quiet
option is to suppress verbose progress output. The purpose of the
--keep-going option is to support long-running and non-interactive
verification of one or more repositories. In which case instant progress
output is of low importance but a summary of errors detected is invaluable
so it makes little sense to suppress the summary.

* subversion/svnadmin/svnadmin.c
  (subcommand_verify): Display the --keep-going error summary regardless
   of whetehr the --quiet option was passed.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1837790 13f79535-47bb-0310-9956-ffa450edef68